refactor/about-page
